Changed all delegates to multicast type and moved client creation code
All delegates are now multicast which means C++ developers will be using AddDynamic() to bind to delegates instead of BindDynamic(), this change was made to keep code consistent and more flexible. CreateDefaultClient was moved from NakamaLibrary in the NakamaBlueprints module to the NakamaClient in the NakamaUnreal module, as it is shared between C++ and Blueprints, demo blueprint controller was also updated to use the function that was moved.